2003-10-29  Al Riddoch  <alriddoch@zepler.org>
	* common/const.cpp, common/const.h, common/random.h,
	  physics/BBox.h, physics/Collision.cpp, physics/Collision.h,
	  physics/Quaternion.cpp, physics/Quaternion.h, physics/Vector3D.h,
	  rulesets/Movement.cpp, rulesets/Plant.cpp, rulesets/Python_API.cpp,
	  tests/collisiontest.cpp: Update variables relating to geometry
	  from double to float as this is the type used internall by
	  WFMath.

	* physics/Vector3D.cpp, physics/Quaternion.cpp, physics/Quaternion.h:
	  Remove legacy code.

	* server/CommPSQLSocket.cpp: Reformatting.

+++ common/const.cpp
@@ -31,22 +31,22 @@ namespace consts {
   int debug_level = 0;
 
   // Scale factor for time. The may become non-constant
-  double time_multiplier = 1.0;
+  float time_multiplier = 1.0;
   // In Real time how many seconds between ticks;
-  double basic_tick = time_multiplier * 3.0;
+  float basic_tick = time_multiplier * 3.0;
   // Highest possible person velocity
-  double base_velocity_coefficient = 5.0;
-  double base_velocity = base_velocity_coefficient/time_multiplier;
+  float base_velocity_coefficient = 5.0;
+  float base_velocity = base_velocity_coefficient/time_multiplier;
   // Square versions of above, to avoid square roots in calculations
-  double square_basic_tick = basic_tick * basic_tick;
-  double square_base_velocity = base_velocity * base_velocity;
+  float square_basic_tick = basic_tick * basic_tick;
+  float square_base_velocity = base_velocity * base_velocity;
 
   // Are ranges for broadcast enabled.
   bool enable_ranges = true;
   // Range of broadcast sight ops
-  double sight_range = 26.0;
+  float sight_range = 26.0;
   // Range of broadcast sound ops
-  double hearing_range = 13.0;
+  float hearing_range = 13.0;
   // Can things be omnipresent
   bool enable_omnipresence = false;
