+++ physics/Collision.h
@@ -17,12 +17,12 @@ typedef std::multimap<int, WFMath::Vector<3> > NormalSet;
 // Returns true if p is infront of plane before collision
 bool getCollisionTime(const Vector3D & p,     // Position of point
                       const Vector3D & u,     // Velocity of point
-                      // double point_time,   // Time since position set
+                      // float point_time,   // Time since position set
                       const Vector3D & l,     // Position on plane
                       const Vector3D & n,     // Plane normal
                       const Vector3D & v,     // Velocity of plane
-                      // double plane_time,   // Time since position set
-                      double & time);         // Returned collision time
+                      // float plane_time,   // Time since position set
+                      float & time);         // Returned collision time
 
 // Predict collision between two sets of meshes, defined by vertices
 // and surface normals
@@ -33,14 +33,14 @@ bool predictCollision(const CoordList & l,    // Vertices of this mesh
                       const CoordList & o,    // Vertices of other mesh
                       const NormalSet & on,   // Normals of other mesh
                       const Vector3D & v,     // Velocity of other mesh
-                      double & time,          // Returned time to collision
+                      float & time,          // Returned time to collision
                       Vector3D & normal);     // Returned collision normal
 
 // Predict collision between 2 entity locations
 // Returns whether the collision will occur
 bool predictCollision(const Location & l,     // Location data of this object
                       const Location & o,     // Location data of other object
-                      double & time,          // Returned time to collision
+                      float & time,          // Returned time to collision
                       Vector3D & normal);     // Returned collision normal
 
 #endif // PHYSICS_COLLISION_H
